refactor(bindings/java): split different classes
