+++ bindings/java/src/main/java/org/apache/opendal/Operator.java
@@ -17,7 +17,6 @@
  * under the License.
  */
 
-
 package org.apache.opendal;
 
 import java.util.Map;
